Axels_Better_Colorization_Machine/requirements.txt
# Standard PyTorch
torch
torchvision

# Microsoft's hardware bridge for AMD GPUs on Windows
#If you have an AMD GPU and are on Windows, you can install torch-directml to leverage DirectML for acceleration. If you're on a different platform or don't have an AMD GPU, you can skip this.
torch-directml

# --- Image Processing & Math ---
numpy
Pillow
scikit-image

# --- Visualization ---
matplotlib